moved controller kill test to standard charm tests
